+++ changelog.rst
@@ -5,6 +5,8 @@ Release History
 0.3.3 (dev)
 +++++++++++
 
+* [changed] Analysis tools restructured
+
 0.3.2 (2015-02-13)
 ++++++++++++++++++